backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Swippter Backend Codebase"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "bcrypt>=5.0.0",
    "celery>=5.6.0",
    "coverage>=7.12.0",
    "django>=5.2.8",
    "django-cors-headers>=4.9.0",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.5.1",
    "drf-spectacular>=0.29.0",
    "gunicorn>=23.0.0",
    "mysqlclient>=2.2.7",
    "pydantic[email]>=2.12.5",
    "python-decouple>=3.8",
    "redis>=7.1.0",
]
